(chore) move list stories to list/stories folder
